GG Bypasser
This is meant to bypass Guardian on your chromebook, instructions to modify to your own specific use case is detailed in the payload as a REM comment (and the only REM comment, find it via ctrl + f).
IMPORTANT: Log into your chromebook after a power wash and IMMEDIATELY disconnect wifi, then plug in the USB loaded with the payload for this to work. If done correctly, Guardian should not have installed as an extension on your browser.
Note: The default network it will bypass is the first network in your list, so if you don't have privelleges to modify it's DNS, this payload will fail. This is also highly experimental and may not work 100% of the time.